Add custom logger module
The custom logger replaces directly calling winston
for logging.
