+++ lib/Facebook.js
@@ -1,9 +1,9 @@
 /** @module facebook */
-const winston = require('winston');
 const Botkit = require('botkit');
 
 const Config = require('../config/Config.js');
 const Services = require('./Services.js');
+const Logger = require('./Logger.js');
 
 const controller = Botkit.facebookbot({
   debug: true,
@@ -70,7 +70,7 @@ function start(bot, cb) {
     cb(err, webserver);
 
     controller.createWebhookEndpoints(webserver, bot, () => {
-      winston.log('info', 'Your borq bot is online');
+      Logger.log('info', 'Your borq bot is online');
     });
 
     if (Config.sentryDSN) {
