+++ lib/HTTP.js
@@ -1,49 +1,11 @@
 /** @module http */
 const fetch = require('node-fetch');
-const raven = require('raven');
-const winston = require('winston');
 
 const Config = require('../config/Config.js');
+const Logger = require('./Logger.js');
 
 const defaultMaxRetries = /prod\w*/i.test(Config.environment) ? 3 : 0;
 const defaultBackoff = 60000;
-
-/**
- * Configure Sentry using the env vars pased if in production
- * @param {string} sentryDSN The url given by Sentry tied to your account
- * @param {string} environment dev, testing, peoduction etc.
- * @param {string} loggingLevel error, info, warn etc.
- * @return {object} return raven object if in production
- */
-function setupSentry(sentryDSN, environment, loggingLevel) {
-  raven
-    .config(sentryDSN, {
-      logger: loggingLevel,
-      environment,
-    })
-    .install();
-
-  if (Config.sentryDSN) {
-    return raven;
-  }
-  return null;
-}
-
-const sentry = Config.sentryDSN
-  ? setupSentry(Config.sentryDSN, Config.environment, 'error')
-  : undefined;
-
-/**
- * A function to send rejected promise errors to both winston and sentry
- * @param {string} text the promise rejection error text
- */
-function genericCatchRejectedPromise(text) {
-  const message = 'Promise rejected: ' + text;
-  winston.log('error', message);
-  if (Config.environment === 'production') {
-    sentry.captureMessage(message);
-  }
-}
 
 /**
  * Extract the response body from a response.
@@ -77,7 +39,7 @@ function processResponse(response, payload) {
    * @param {string} r - response text to log
    */
   function log(r) {
-    winston.log(
+    Logger.log(
       logLevel,
       `{
 Status: ${status}
@@ -144,11 +106,7 @@ function request(
         }
       }
     })
-    .catch((reason) => genericCatchRejectedPromise(rejectionText + reason));
+    .catch((reason) => Logger.logRejectedPromise(rejectionText + reason));
 }
 
-module.exports = {
-  request,
-  sentry,
-  genericCatchRejectedPromise,
-};
+module.exports = request;
