+++ lib/Translations.js
@@ -1,10 +1,10 @@
 /** @module translations */
 const fs = require('fs');
-const winston = require('winston');
 const i18next = require('i18next');
 const Backend = require('i18next-node-fs-backend');
 
 const config = require('../config/Config.js');
+const Logger = require('./Logger.js');
 
 /**
  * Generate namespaces needed by i18next from the files in the translations
@@ -43,13 +43,13 @@ function translateiUsing18next() {
 
     i18next.use(Backend).init(i18nextOptions, (err, t) => {
       if (err) {
-        winston.log('error', `Error loading translations: ${t}`, err);
+        Logger.log('error', `Error loading translations: ${t}`, err);
       } else {
-        winston.log('info', 'Translations loaded successfully');
+        Logger.log('info', 'Translations loaded successfully');
       }
     });
   } else {
-    winston.log('info', 'Not using translations');
+    Logger.log('info', 'Not using translations');
   }
 
   /**
