+++ lib/services/RapidPro.js
@@ -1,8 +1,7 @@
 /** @module services.rapidpro */
-const winston = require('winston');
-
 const http = require('../../lib/HTTP.js');
 const config = require('../../config/Config.js');
+const Logger = require('./Logger.js');
 
 const rapidProBaseURL = 'https://rapidpro.ona.io/api/v2/';
 const contactsEndPoint = `${rapidProBaseURL}contacts.json`;
@@ -55,7 +54,7 @@ function updateContact({uuid, urn}, payload, baseURL = contactsEndPoint) {
 function deleteContact(messengerId, groups, url = contactsEndPoint) {
   return postRapidPro({groups}, `${url}?urn=facebook:${messengerId}`)
     .then(({uuid, groups}) => {
-      winston.log(
+      Logger.log(
         'info',
         `Deleted user ${messengerId} ` + ` group(s) ${JSON.stringify(groups)}.`
       );
